serverless_backend/Module-5Machine Learning/feedbackScorePredictor_GoogleCloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-aiplatform
protobuf
nltk
google-cloud-firestore